pyproject.toml
[tool.poetry]
name = "dense-prediction-transformer"
version = "0.1.0"
description = ""
authors = ["Maruf Ahmed <maruf@technyx.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
torch = "^2.6.0"
numpy = "^2.2.3"
opencv-python = "^4.11.0.86"
transformers = "^4.48.3"
pillow = "^11.1.0"
tk-tools = "^0.17.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
